ARM: dts: imx6ul-phycore: Add state framework
Add the state framework with bootchooser nodes on EEPROMs and enable it
only on the full featured boards with NAND as the low cost boards have not
enough space on the NAND for a redundant boot system. Also boards with
eMMC are excluded from the state framework as we do not support RAUC
with eMMC at this point.

Signed-off-by: Stefan Riedmueller <s.riedmueller@phytec.de>
Signed-off-by: Christian Hemp <c.hemp@phytec.de>