mtd: ubi: Max out default wear-leveling threshold
Due to the recursive implementation of ubi wear-leveling in the barebox,
a large amount of wear-leveling can lead to a stack overflow.

This was observed during extensive ubi stress tests with periodic power
cycles.

As the kernel is perfectly capable of handling wear-leveling we can
increase the wear-leveling threshold in the barebox to leave it to the
kernel. To minimize the chance of wear-leveling in the barebox we max
out the threshold.

Of course in general the recursive part of the wear-leveling worker
should be reconsidered to reduce the risk of stack overflows even more.

Signed-off-by: Stefan Riedmueller <s.riedmueller@phytec.de>
Signed-off-by: Christian Hemp <c.hemp@phytec.de>