mtd: nand_bbt: Use mtd block_markbad function
Using the nand_chip block_markbad function might not be able to mark a
block as bad directly. This is the case for instance with the mxs nand
driver as its hooks are only called with the mtd block_markbad function.
So make use of the latter one.

This includes invalidating the td-pages entry before actually marking
the block bad and updating the in-memory BBT afterwards in case it was
not already updated by the block_markbad function.

Signed-off-by: Stefan Riedmueller <s.riedmueller@phytec.de>
Signed-off-by: Christian Hemp <c.hemp@phytec.de>