fs: Use explicit mount point for cdev default automount entries
Automounting devices without explicitly setting the mountpoint can lead
to a system failure if the device is already mounted to a different
mount point than the default one. In this cause the system is caught in
an infinite loop spamming:

mounted /dev/mmc0.0 on /mnt/mmc
mounted /dev/mmc0.0 on /mnt/mmc
mounted /dev/mmc0.0 on /mnt/mmc
...

and not responding. Avoid this by simply setting the mount point
explicitly for cdev default automount entries.

Signed-off-by: Stefan Riedmueller <s.riedmueller@phytec.de>
Signed-off-by: Christian Hemp <c.hemp@phytec.de>